Add Messenger for message emission; split range metadata into _ranges
Messenger wraps ConversionResultsBuilder with severity-named methods and
accepts a CellRangeMetadata directly as the Excel reference; all 81
addMessage call sites now go through it.

CellRangeMetadata/XbrlConceptCellRangeMetadata move to _ranges.py along
with the iteration helpers (absorbing _cell_iteration.py), leaving
_bindings.py with only the binding aggregates.
