+++ tests/unitTests/xlsx_template_reader/test_ranges.py
@@ -5,7 +5,7 @@ from openpyxl import Workbook
 from openpyxl.worksheet.cell_range import CellRange
 from openpyxl.worksheet.worksheet import Worksheet
 
-from mireport.xlsx_template_reader._cell_iteration import (
+from mireport.xlsx_template_reader._ranges import (
     _CellRangeDimensions,
     getEffectiveCellRangeDimensions,
 )
